Change DlgReportView file names.
