+++ src/Gui/DlgSettingsReportView.ui
